+++ src/plugins/relay/api/remote/relay-remote-event.c
@@ -1171,6 +1171,16 @@ RELAY_REMOTE_EVENT_CALLBACK(upgrade)
 }
 
 /*
+ * Callback for event "quit".
+ */
+
+RELAY_REMOTE_EVENT_CALLBACK(quit)
+{
+    relay_remote_network_disconnect (event->remote);
+    return WEECHAT_RC_OK;
+}
+
+/*
  * Synchronizes with remote.
  */
 
@@ -1227,6 +1237,7 @@ relay_remote_event_recv (struct t_relay_remote *remote, const char *data)
         { "nicklist_group_*", &relay_remote_event_cb_nick_group },
         { "nicklist_nick_*", &relay_remote_event_cb_nick },
         { "upgrade*", &relay_remote_event_cb_upgrade },
+        { "quit", &relay_remote_event_cb_quit },
         { NULL, NULL },
     };
     t_relay_remote_event_func *callback;